+++ src/main/java/org/truffleruby/core/CoreLibrary.java
@@ -1438,7 +1438,6 @@ public class CoreLibrary {
             "/core/random.rb",
             "/core/truffle/io_operations.rb",
             "/core/truffle/kernel_operations.rb",
-            "/core/truffle/thread_operations.rb",
             "/core/thread.rb",
             "/core/true.rb",
             "/core/type.rb",